Duende IdentityServer

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -51,12 +52,13 @@ public class AuctionController : ControllerBase
         return _mapper.Map<AuctionDTO>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDTO>> CreateAuction(CreateAuctionDTO auctionDTO)
     {
         var auction = _mapper.Map<Auction>(auctionDTO);
-        //TODO: add current user as seller
-        auction.Seller = "test";
+
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
 
@@ -71,6 +73,7 @@ public class AuctionController : ControllerBase
         return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDTO updateAuctionDTO)
     {
@@ -79,7 +82,7 @@ public class AuctionController : ControllerBase
 
         if (auction is null) return NotFound();
 
-        //TODO: Check seller username
+        if(auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDTO.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDTO.Model ?? auction.Item.Model;
@@ -97,6 +100,7 @@ public class AuctionController : ControllerBase
 
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -104,7 +108,7 @@ public class AuctionController : ControllerBase
 
         if (auction is null) return NotFound();
 
-        // TODO: check seller username
+        if(auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
